Test running parse csv on entire file, all old tests are in tests_backup
