+++ poll_sample.csv
@@ -1,0 +1,10 @@
+Timestamp:first,Timestamp:last,Are you ready to report an idling vehicle?,Are you ready to report an idling vehicle? (option value),"What is your location? Be specific! If you have a timer, press start!",What agency owns the vehicle you see idling?,What agency owns the vehicle you see idling? (option value),What agency or company?,What is the vehicle ID #? Usually on the back or side of the vehicle.  ,What is the license plate number? If you don't see it that's ok.,How many minutes have you observed the vehicle idling? Stop your timer! Be specific!,Please let us know of the general situation (i.e. is the vehicle unoccupied).,"Lastly, can you take a photo of the vehicle and text it back to us? Make sure plate or ID # is visible.","Lastly, can you take a photo of the vehicle and text it back to us? Make sure plate or ID # is visible. (media url)",Area Code,Phone Prefix
+10/2/2015 9:40,10/2/2015 9:43,Y,Yes,15 & chestnut ,G,Other,Abbonizio transfer ,250,AG-26081,"6, but idling before and after I arrived and left ","Driver in truck, possibly eating, windows partially down",,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/255089/IMG_2937.jpg,19174469999,744
+10/2/2015 9:35,10/2/2015 9:40,Y,Yes,Poplar & n American ,G,Other,Verizon ,220530591,YSK-9618,"10, but idling before and after I arrived and left ","Driver sleeping in truck, no work being done.",Y,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/255080/IMG_3656.jpg,19174469999,744
+10/2/2015 9:29,10/2/2015 9:35,Y,Yes,Broad & arch ,E,Streets, ,105014,MG-0512E,"10, but idling before and after I arrived and left ","Driver in truck talking on phone, window partially down, no work being done",,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/255071/IMG_5074.jpg,19174469999,744
+10/1/2015 18:17,10/2/2015 9:29,Y,Yes,Y,G,Other,Xx,Y,Y,Y,Y,Y,,19174469999,744
+10/1/2015 11:12,10/1/2015 18:17,Y,Yes,City hall ,E,Streets, ,995035,75559-MG,"20, but idling before and after I arrived and left ","No driver, passenger sitting in vehicle, no work being done, door unlocked.",,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/254752/IMG_8482.jpg,19174469999,744
+10/1/2015 9:27,10/1/2015 11:12,Y,Yes,Broad & arch ,E,Streets, ,85237,MG-7677D,"7, but idling before and after I arrived and left ","Driver and passenger sitting in truck texting and talking, no work being done, windows down slightly",,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/254671/IMG_2909.jpg,19174469999,744
+9/30/2015 9:21,10/1/2015 9:27,Y,Yes,Spring garden & 4,C,PWD, ,60063,Didn't see it ,3,"Driver came back from getting food, turned on engine and then set up his steering wheel like a table. After a minute I went over to take a picture, he saw me, tried to take my picture, then drove off. @",,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/254451/IMG_8991.jpg,19174469999,744
+9/29/2015 12:31,9/30/2015 9:21,Y,Yes,18 & walnut,G,Other,Generation 3 electric,106,YYY-9328,"15, but idling before and after I arrived and left ",Employees sitting in van eating lunch. No work being done with vehicle. Said they wanted it on to stay cool.,,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/254209/IMG_2900.jpg,19174469999,744
+9/29/2015 16:55,9/29/2015 16:58,Yes,Yes,Walnut at 36th,E,Streets, ,126036,Mg9769f,3 mins. Idling when I arrived.,2 employees waiting in idling vehicle with passenger window open. After 3 mins of observation a third employee went into vehicle. Drove away a minute later.,,https://textizen-attachments.s3.amazonaws.com/uploads/response/photo_attachment/254336/IMG_3738.jpg,12154501504,545
